SECURITY-894 Improve SimpleRoleGroup Scalability
During load testing we found out that SimpleRoleGroup does not scale
well to large numbers of roles (200 in our case). The reason for this
is that it uses ArrayList#contains with scales linearly with the size
of the collection. Since the method is called in a loop the scalability
is actually quadratic.

This commit includes the following changes:

 * change SimpleRoleGroup#roles from ArrayList to HashSet
 * change RoleGroup#getRoles and #addAll from List to Collection
 * add a test that verifies old serialized objects can still be loaded
 * add a test that verifies serialization still works

A quick quick with the WildFly source suggests that the change is
probably source compatible in WildFly but not binary compatible.

See the following mailing list discussion [1].

 [1] http://lists.jboss.org/pipermail/wildfly-dev/2015-June/004033.html

Issue: SECURITY-894
https://issues.jboss.org/browse/SECURITY-894